requirements.txt
fastapi
uvicorn
requests
python-multipart
aiofiles
httpx
python-dotenv
